First Upload
Getting the working code up here. It's certainly not perfect, needs more options. This though, is the hard work concerning AI. There is grunt work for getting codenames to work, maybe even a graphical interface but thats for other, different type of knowledge person. Like Matt. But it already exists.